course list image concept test
### Why

Proof of Concept using temporary Icon to trigger Course List Modals

Style updates for comments

### How

Added a Grade9 icon from stock photos as a placeholder for the icons
that are not yet available from marketing.

Removed italics. Added a timeline node as a placeholder for upcoming
content . Corrected .scss for courses section to put + sign back on
picture when hovering.